package rename: remove unused modules
